requirements.txt
aiogram==3.4.1
opencv-python-headless
ffmpeg-python
